# backend/packages/api-spec/openapi.yaml
openapi: 3.1.0
info:
  title: Api
  version: 1.0.0
  description: AAAgronegócio Platform API - Farm Listings Marketplace & SaaS
servers:
  - url: /api
    description: Base API path
...
paths:
...
  /listings:
    get:
      operationId: getListings
      tags: [listings]
      summary: Get published listings with filters and pagination
      parameters:
        - name: page
          in: query
          schema:
            type: integer
            minimum: 1
            default: 1
        - name: limit
          in: query
          schema:
            type: integer
            minimum: 1
            maximum: 50
            default: 12
        - name: search
          in: query
          schema:
            type: string
        - name: state
          in: query
          schema:
            type: string
        - name: city
          in: query
          schema:
            type: string
        - name: country
          in: query
          schema:
            type: string
        - name: minPrice
          in: query
          schema:
            type: number
        - name: maxPrice
          in: query
          schema:
            type: number
        - name: minArea
          in: query
          schema:
            type: number
        - name: maxArea
          in: query
          schema:
            type: number
        - name: propertyType
          in: query
          schema:
            type: string
        - name: aptitude
          in: query
          schema:
            type: string
          description: Texto livre; busca em descrição e campos de atividade da propriedade
        - name: hasWater
          in: query
          schema:
            type: boolean
        - name: hasIrrigation
          in: query
          schema:
            type: boolean
        - name: sortBy
          in: query
          schema:
            type: string
            enum: [price_asc, price_desc, area_asc, area_desc, newest, featured]
            default: featured
        - name: lat
          in: query
          schema:
            type: number
        - name: lng
          in: query
          schema:
            type: number
        - name: radius
          in: query
          schema:
            type: number
      responses:
        "200":
          description: Listings list
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ListingsPage"
...
  /admin/listings:
    get:
      operationId: adminGetListings
      tags: [admin]
      summary: Get all listings (admin view)
      security:
        - cookieAuth: []
      parameters:
        - name: page
          in: query
          schema:
            type: integer
            default: 1
        - name: limit
          in: query
          schema:
            type: integer
            default: 20
        - name: status
          in: query
          schema:
            type: string
        - name: moderationStatus
          in: query
          schema:
            type: string
      responses:
        "200":
          description: All listings
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/AdminListingsPage"
  /admin/listings/{id}/approve:
    post:
      operationId: approveListing
      tags: [admin]
      summary: Approve a listing for publication
      security:
        - cookieAuth: []
      parameters:
        - name: id
          in: path
          required: true
          schema:
            type: string
      responses:
        "200":
          description: Approved
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ListingDetail"
  /admin/listings/{id}/reject:
    post:
      operationId: rejectListing
      tags: [admin]
      summary: Reject a listing with reason
      security:
        - cookieAuth: []
      parameters:
        - name: id
          in: path
          required: true
          schema:
            type: string
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/RejectListingRequest"
      responses:
        "200":
          description: Rejected
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ListingDetail"
  /admin/listings/{id}/feature:
    post:
      operationId: featureListing
      tags: [admin]
      summary: Toggle featured status on a listing
      security:
        - cookieAuth: []
      parameters:
        - name: id
          in: path
          required: true
          schema:
            type: string
      requestBody:
        required: true
        content:
          application/json:
            schema:
              type: object
              properties:
                featured:
                  type: boolean
              required:
                - featured
      responses:
        "200":
          description: Feature status updated
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ListingDetail"
  /admin/users:
    get:
      operationId: adminGetUsers
      tags: [admin]
      summary: Get all users
      security:
        - cookieAuth: []
      parameters:
        - name: page
          in: query
          schema:
            type: integer
            default: 1
        - name: limit
          in: query
          schema:
            type: integer
            default: 20
        - name: role
          in: query
          schema:
            type: string
      responses:
        "200":
          description: Users list
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/UsersPage"
  /admin/users/{id}:
    put:
      operationId: adminUpdateUser
      tags: [admin]
      summary: Update user (role, status, premium)
      security:
        - cookieAuth: []
      parameters:
        - name: id
          in: path
          required: true
          schema:
            type: string
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/AdminUpdateUserRequest"
      responses:
        "200":
          description: User updated
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/UserProfile"
  /admin/dashboard:
    get:
      operationId: getAdminDashboard
      tags: [admin]
      summary: Get admin dashboard summary
      security:
        - cookieAuth: []
      responses:
        "200":
          description: Admin dashboard data
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/AdminDashboard"
  /admin/analytics:
    get:
      operationId: getAdminAnalytics
      tags: [admin]
      summary: Get platform-wide analytics
      security:
        - cookieAuth: []
      parameters:
        - name: period
          in: query
          schema:
            type: string
            enum: [7d, 30d, 90d, 1y]
            default: 30d
      responses:
        "200":
          description: Analytics data
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/PlatformAnalytics"

  # ─── AI FEATURES ──────────────────────────────────────────────────────────────
...
components:
  securitySchemes:
    cookieAuth:
      type: apiKey
      in: cookie
      name: access_token
  schemas:
...
    UserProfile:
      type: object
      properties:
        id:
          type: string
        name:
          type: string
        email:
          type: string
        role:
          type: string
          enum: [admin, seller, buyer]
        phone:
          type: string
          nullable: true
        avatarUrl:
          type: string
          nullable: true
        isPremium:
          type: boolean
        isActive:
          type: boolean
        createdAt:
          type: string
          format: date-time
      required:
        - id
        - name
        - email
        - role
        - isPremium
        - isActive
        - createdAt
...
    ListingSummary:
      type: object
      properties:
        id:
          type: string
        slug:
          type: string
        title:
          type: string
        price:
          type: number
          nullable: true
        negotiable:
          type: boolean
        totalArea:
          type: number
          nullable: true
        areaUnit:
          type: string
          nullable: true
        pricePerHectare:
          type: number
          nullable: true
        state:
          type: string
          nullable: true
        city:
          type: string
          nullable: true
        country:
          type: string
          nullable: true
        propertyType:
          type: string
          nullable: true
        isFeatured:
          type: boolean
        isPremium:
          type: boolean
        coverImageUrl:
          type: string
          nullable: true
        thumbnailUrls:
          type: array
          items:
            type: string
        viewCount:
          type: integer
        favoriteCount:
          type: integer
        createdAt:
          type: string
          format: date-time
      required:
        - id
        - slug
        - title
        - negotiable
        - isFeatured
        - isPremium
        - thumbnailUrls
        - viewCount
        - favoriteCount
        - createdAt
    ListingDetail:
      type: object
      properties:
        id:
          type: string
        slug:
          type: string
        title:
          type: string
        description:
          type: string
          nullable: true
        price:
          type: number
          nullable: true
        negotiable:
          type: boolean
        status:
          type: string
          enum: [draft, pending_review, published, rejected, archived]
        moderationStatus:
          type: string
          enum: [pending, approved, rejected]
        moderationNote:
          type: string
          nullable: true
        isFeatured:
          type: boolean
        isPremium:
          type: boolean

        country:
          type: string
          nullable: true
        state:
          type: string
          nullable: true
        city:
          type: string
          nullable: true
        region:
          type: string
          nullable: true
        latitude:
          type: number
          nullable: true
        longitude:
          type: number
          nullable: true
        addressDetails:
          type: string
          nullable: true

        totalArea:
          type: number
          nullable: true
        productiveArea:
          type: number
          nullable: true
        reserveArea:
          type: number
          nullable: true
        areaUnit:
          type: string
          enum: [hectares, alqueires, acres]
          nullable: true
        pricePerHectare:
          type: number
          nullable: true
        pricePerAlqueire:
          type: number
          nullable: true

        propertyType:
          type: string
          nullable: true
        aptidaoAgricola:
          type: boolean
        aptidaoPecuaria:
          type: boolean
        aptidaoMista:
          type: boolean
        hasIrrigation:
          type: boolean
        hasRiver:
          type: boolean
        hasLake:
          type: boolean
        hasAcude:
          type: boolean
        hasWell:
          type: boolean
        hasSpring:
          type: boolean

        infraMainHouse:
          type: boolean
        infraWorkerHouses:
          type: boolean
        infraWarehouse:
          type: boolean
        infraBarn:
          type: boolean
        infraCorral:
          type: boolean
        infraFencing:
          type: boolean
        infraElectricity:
          type: boolean
        infraInternet:
          type: boolean
        infraMachineryShed:
          type: boolean
        infraGrainSilo:
          type: boolean
        infraDairyStructure:
          type: boolean
        infraIrrigationSystem:
          type: boolean
        infraPavedAccess:
          type: boolean
        infraInternalRoads:
          type: boolean
        infraAirstrip:
          type: boolean
        infraOffice:
          type: boolean
        infraSolarPower:
          type: boolean
        infraSecuritySystem:
          type: boolean

        hasCattle:
          type: boolean
        cattleCount:
          type: integer
          nullable: true
        hasHorses:
          type: boolean
        otherAnimals:
          type: string
          nullable: true
        plantedCrops:
          type: string
          nullable: true
        cropDetails:
          type: string
          nullable: true
        productiveActivities:
          type: string
          nullable: true

        bedroomCount:
          type: integer
          nullable: true
        bathroomCount:
          type: integer
          nullable: true
        suiteCount:
          type: integer
          nullable: true
        hasKitchen:
          type: boolean
        hasGarage:
          type: boolean
        hasLeisureArea:
          type: boolean

        deedStatus:
          type: string
          nullable: true
        environmentalLicense:
          type: boolean
        carRegistryInfo:
          type: string
          nullable: true

        media:
          type: array
          items:
            $ref: "#/components/schemas/MediaItem"

        seller:
          $ref: "#/components/schemas/SellerPublicInfo"

        viewCount:
          type: integer
        favoriteCount:
          type: integer
        interestCount:
          type: integer

        createdAt:
          type: string
          format: date-time
        updatedAt:
          type: string
          format: date-time
        publishedAt:
          type: string
          format: date-time
          nullable: true
      required:
        - id
        - slug
        - title
        - negotiable
        - status
        - moderationStatus
        - isFeatured
        - isPremium
        - aptidaoAgricola
        - aptidaoPecuaria
        - aptidaoMista
        - hasIrrigation
        - hasRiver
        - hasLake
        - hasAcude
        - hasWell
        - hasSpring
        - infraMainHouse
        - infraWorkerHouses
        - infraWarehouse
        - infraBarn
        - infraCorral
        - infraFencing
        - infraElectricity
        - infraInternet
        - infraMachineryShed
        - infraGrainSilo
        - infraDairyStructure
        - infraIrrigationSystem
        - infraPavedAccess
        - infraInternalRoads
        - infraAirstrip
        - infraOffice
        - infraSolarPower
        - infraSecuritySystem
        - hasCattle
        - hasHorses
        - hasKitchen
        - hasGarage
        - hasLeisureArea
        - environmentalLicense
        - media
        - viewCount
        - favoriteCount
        - interestCount
        - createdAt
        - updatedAt
...
    ListingsPage:
      type: object
      properties:
        listings:
          type: array
          items:
            $ref: "#/components/schemas/ListingSummary"
        total:
          type: integer
        page:
          type: integer
        limit:
          type: integer
        totalPages:
          type: integer
      required:
        - listings
        - total
        - page
        - limit
        - totalPages
...
    AdminDashboard:
      type: object
      properties:
        totalListings:
          type: integer
        pendingApproval:
          type: integer
        publishedListings:
          type: integer
        totalUsers:
          type: integer
        totalSellers:
          type: integer
        totalBuyers:
          type: integer
        premiumSellers:
          type: integer
        totalViews:
          type: integer
        totalInterests:
          type: integer
        recentListings:
          type: array
          items:
            $ref: "#/components/schemas/ListingSummary"
        listingsByStatus:
          type: array
          items:
            $ref: "#/components/schemas/StatusMetric"
        viewsByDay:
          type: array
          items:
            $ref: "#/components/schemas/DayMetric"
      required:
        - totalListings
        - pendingApproval
        - publishedListings
        - totalUsers
        - totalSellers
        - totalBuyers
        - premiumSellers
        - totalViews
        - totalInterests
        - recentListings
        - listingsByStatus
        - viewsByDay
...
    PlatformAnalytics:
      type: object
      properties:
        period:
          type: string
        totalViews:
          type: integer
        uniqueVisitors:
          type: integer
        totalInterests:
          type: integer
        conversionRate:
          type: number
        avgPriceByState:
          type: array
          items:
            $ref: "#/components/schemas/StateMetric"
        viewsByDay:
          type: array
          items:
            $ref: "#/components/schemas/DayMetric"
        topListings:
          type: array
          items:
            $ref: "#/components/schemas/ListingSummary"
        interestsByType:
          type: array
          items:
            $ref: "#/components/schemas/InterestTypeMetric"
      required:
        - period
        - totalViews
        - uniqueVisitors
        - totalInterests
        - conversionRate
        - avgPriceByState
        - viewsByDay
        - topListings
        - interestsByType
...
    AdminListingsPage:
      type: object
      properties:
        listings:
          type: array
          items:
            $ref: "#/components/schemas/ListingDetail"
        total:
          type: integer
        page:
          type: integer
        limit:
          type: integer
        totalPages:
          type: integer
      required:
        - listings
        - total
        - page
        - limit
        - totalPages
    UsersPage:
      type: object
      properties:
        users:
          type: array
          items:
            $ref: "#/components/schemas/UserProfile"
        total:
          type: integer
        page:
          type: integer
        limit:
          type: integer
        totalPages:
          type: integer
      required:
        - users
        - total
        - page
        - limit
        - totalPages
    AdminUpdateUserRequest:
      type: object
      properties:
        role:
          type: string
          enum: [admin, seller, buyer]
        isPremium:
          type: boolean
        isActive:
          type: boolean
    RejectListingRequest:
      type: object
      properties:
        reason:
          type: string
          minLength: 5
      required:
        - reason

    # ─── AI ──────────────────────────────────────────────────────────────────────
...